wip: partial state update by adding ssz_rs to ream
