+++ consensus/Cargo.toml
@@ -9,3 +9,5 @@ ethereum_ssz.workspace = true
 ream-consensus.workspace = true
 tree_hash.workspace = true
 anyhow = "1.0"
+ssz_rs.workspace = true
+ssz_rs_derive.workspace = true
