+++ consensus/src/merkleized_validator.rs
@@ -1,0 +1,6 @@
+use ream_consensus::validator::Validator;
+
+pub struct MerklelizedValidator {
+  validator: Validator,
+  proof: u32,
+}
