+++ methods/guest/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 
 [dependencies]
 consensus = { path = "../../consensus" }
-ream-consensus = { git = "https://github.com/ReamLabs/ream", package = "ream-consensus" }
+ream-consensus = { git = "git@github.com:unnawut/ream.git", branch = "ssz_rs", package = "ream-consensus" }
 risc0-zkvm = { version = "1.2.0", default-features = false, features = ['std'] }
 serde = { version = "1.0", features = ["derive"] }
