+++ methods/guest/src/main.rs
@@ -1,6 +1,6 @@
 use risc0_zkvm::guest::env;
 
-use consensus;
+use consensus::beacon_state;
 use ream_consensus::deneb::beacon_block::BeaconBlock;
 use ream_consensus::beacon_block_header::BeaconBlockHeader;
 
@@ -12,7 +12,7 @@ fn main() {
     let proposer_index: u64 = env::read();
     let block: BeaconBlock = env::read();
 
-    let new_block_header = consensus::partials_process_block_header(
+    let new_block_header = beacon_state::process_block_header(
             slot,
             &latest_block_header,
             proposer_index,
